# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * AI Lesson Planner API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

/**
 * Desired lesson depth. Quick is a concise overview, standard is a balanced session, and deep is a detailed lesson.
 */
export type LessonPlanInputDepth = typeof LessonPlanInputDepth[keyof typeof LessonPlanInputDepth];


export const LessonPlanInputDepth = {
  quick: 'quick',
  standard: 'standard',
  deep: 'deep',
} as const;

export interface LessonPlanInput {
  /** @minLength 1 */
  topic: string;
  /** Desired lesson depth. Quick is a concise overview, standard is a balanced session, and deep is a detailed lesson. */
  depth?: LessonPlanInputDepth;
  /** The signed-in user's secret tracking code. */
  secretCode: string;
}

export interface QuizQuestion {
  question: string;
  options: string[];
  /** Index of the correct option (0-based) */
  correctAnswer: number;
  explanation: string;
}

export interface LessonDay {
  id: number;
  dayNumber: number;
  title: string;
  /** Markdown content for the lesson */
  lessonContent: string;
  quiz: QuizQuestion[];
  createdAt: string;
}

export interface LessonPlan {
  id: number;
  topic: string;
  createdAt: string;
  days: LessonDay[];
}

export interface LessonPlanSummary {
  id: number;
  topic: string;
  createdAt: string;
  dayCount: number;
}